refactor: 클래스 분리 및 코드 정리
