+++ src/main/java/http/HttpQueryParams.java
@@ -1,4 +1,4 @@
-package model;
+package http;
 
 import java.util.HashMap;
